revert(sdk-dev): workspace package split was half-baked, revert to single file
Stage 3 turned workspace.py into workspace/ package but the 'split' was
purely cosmetic - the entire 1402-line Workspace class still lived in
workspace/base.py. Net effect:

  Before: agentica/workspace.py (single file, direct)
  After:  agentica/workspace/base.py + __init__.py  (pointless indirection)

This 'fake split' actually made things worse:
- Tests had to patch agentica.workspace.base.AGENTICA_HOME (1 extra segment)
- Added maintenance burden of an extra __init__.py
- Did not deliver any of the advertised benefits (no logical separation)

A real split into memory.py / git_context.py / experience.py / etc. requires
breaking apart a tightly-coupled 55-method class, which is too risky for a
v1.3.x maintenance release.

Reverted:
- workspace/base.py -> workspace.py (single file, as on main)
- workspace/__init__.py deleted
- tests/test_workspace.py: patch paths restored to agentica.workspace.AGENTICA_HOME
- tests/test_skill_lazy_loading.py: import back to agentica.workspace

Future logical split will happen in v2.0 (when we have the scope to do it
properly with mixin decomposition).

Tests: 1349 passed, 0 regression.
Made-with: Cursor
